+++ wgpu-hal/src/metal/adapter.rs
@@ -1438,8 +1438,10 @@ impl super::CapabilitiesQuery {
             max_buffer_size: self.max_buffer_size,
             // No limit, use maxBufferSize.
             max_uniform_buffer_binding_size: self.max_buffer_size,
-            // No limit, use maxBufferSize.
-            max_storage_buffer_binding_size: self.max_buffer_size,
+            // naga bounds-checks use `uint`. Limit to `u32::MAX` if enabled.
+            max_storage_buffer_binding_size: self
+                .max_buffer_size
+                .min(u64::from(u32::MAX) & !(wgt::STORAGE_BINDING_SIZE_ALIGNMENT as u64 - 1)),
             min_uniform_buffer_offset_alignment: self.constant_buffer_offset_alignment,
             // No documented limit. Use 32, which is the lowest allowed value.
             min_storage_buffer_offset_alignment: 32,